hal: update channel coefficient memory allocation
- observing pulse audio crash for multi channel PCM playback.
- memory allocation for coefficients in audio_custom_mtmx_params
  is going wrong
- Fix the crash issue by allocating memory for correct data type

+++ hal/audio_extn/audio_extn.c
@@ -588,9 +588,9 @@ static struct audio_custom_mtmx_params *update_channel_weightage_params(
 
     /*
      * Allocate memory for coefficients in audio_custom_mtmx_params.
-     * Coefficent in audio_custom_mtmx_params is of type uint32_t.
+     * Coefficent in audio_custom_mtmx_params is of type long int.
      */
-    size += sizeof(uint32_t) * ip_channels * op_channels;
+    size += sizeof(long int) * ip_channels * op_channels;
     params = (struct audio_custom_mtmx_params *) calloc(1, size);
 
     if (!params) {
